+++ src/main/java/de/samply/fhir/BundleContext.java
@@ -1,5 +1,6 @@
 package de.samply.fhir;
 
+import ca.uhn.fhir.context.FhirContext;
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Table;
 import de.samply.template.AttributeTemplate;
@@ -12,13 +13,16 @@ public class BundleContext {
 
 
   private ResourceFinder resourceFinder;
+  private BundleValidator bundleValidator;
   private BundleToContainersConverterSession session;
   private Table<String, String, String> idValueAnonymMap = HashBasedTable.create();
 
+
   public BundleContext(Bundle bundle, BundleToContainersConverterSession session,
-      FHIRPathEngine fhirPathEngine) {
+      FHIRPathEngine fhirPathEngine, FhirContext fhirContext) {
     this.resourceFinder = new ResourceFinder(bundle, fhirPathEngine);
     this.session = session;
+    this.bundleValidator = new BundleValidator(fhirContext);
   }
 
   public String fetchAnonym(AttributeTemplate attributeTemplate, String value) {
@@ -32,6 +36,10 @@ public class BundleContext {
 
   public List<Resource> fetchRelatedResources(Resource currentResource, AttributeTemplate attributeTemplate){
     return resourceFinder.fetchRelatedResources(currentResource, attributeTemplate);
+  }
+
+  public boolean validate(Resource resource, AttributeTemplate attributeTemplate) {
+    return bundleValidator.validate(resource, attributeTemplate);
   }